backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.0",
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
